Seperated successful and exception tests of connection into different tests
